requirements.txt
flask==2.0.1
openai==1.3.0
yt-dlp==2023.11.16
pydub==0.25.1
ffmpeg-python==0.2.0
python-dotenv==0.19.0
